Se agrego salto y proyectil
Proyectil_component.tscn:
A scene that creates a bullet that the cat can use, it only goes through the x axis.

Proyectil_component.gd:
This script has the hitbox function for when an enemy hits the bullet and gets hit.

PathFindState.gd:
Made it so the cat can also jump and has gravity, it will fall unless a rigidBody or tilemap is under the cat.

Worldpostre.tscn:
Just a world i created to test the jumping mechanics and the bullets.