Improvement, EnemyRat and EnemyJetRat, Added 2 new enemies
EnemyRat:
In the idle state it goes from left to right until it sees the player, after seeing the player it will use a dash if it can otherwise it will go to the player, when its close enough it weill attack.

EnemyJetRat:
In the idle state this enemy will stay still and has a big area it can detect the player in, after finding the player it will launch itself at the player and begin to fly, if the player parries the attack then the rat will fall to the ground and attack as if it was a normal EnemyRat wihout the dash. If it manages to hit the player in its launch attack it will explode (explosion is yet to be implemented)